+++ include/Refine3D.h
@@ -110,7 +110,7 @@ template<typename real_t> class Refine3D{
     std::vector< std::vector<index_t> > refined_edges(NNodes);
     std::vector< std::vector< DirectedEdge<index_t> > > newVertices(nthreads);
     std::vector< std::vector<real_t> > newCoords(nthreads);
-    std::vector< std::vector<float> > newMetric(nthreads);
+    std::vector< std::vector<double> > newMetric(nthreads);
     std::vector< std::vector<index_t> > newElements(nthreads);
     std::vector<size_t> threadIdx(nthreads), splitCnt(nthreads, 0);
 
@@ -380,7 +380,7 @@ template<typename real_t> class Refine3D{
 
       // Append new coords and metric to the mesh.
       memcpy(&_mesh->_coords[ndims*threadIdx[tid]], &newCoords[tid][0], ndims*splitCnt[tid]*sizeof(real_t));
-      memcpy(&_mesh->metric[msize*threadIdx[tid]], &newMetric[tid][0], msize*splitCnt[tid]*sizeof(float));
+      memcpy(&_mesh->metric[msize*threadIdx[tid]], &newMetric[tid][0], msize*splitCnt[tid]*sizeof(double));
 
       assert(newVertices[tid].size()==splitCnt[tid]);
       for(size_t i=0;i<splitCnt[tid];i++){
@@ -682,7 +682,7 @@ template<typename real_t> class Refine3D{
  private:
 
   void refine_edge(index_t n0, index_t n1, std::vector< DirectedEdge<index_t> > &newVertices,
-      std::vector<real_t> &coords, std::vector<float> &metric){
+      std::vector<real_t> &coords, std::vector<double> &metric){
     if(lnn2gnn[n0]>lnn2gnn[n1]){
       // Needs to be swapped because we want the lesser gnn first.
       index_t tmp_n0=n0;
@@ -695,10 +695,10 @@ template<typename real_t> class Refine3D{
     // Li et al, Comp Methods Appl Mech Engrg 194 (2005) 4915-4950.
     real_t x, m;
     const real_t *x0 = _mesh->get_coords(n0);
-    const float *m0 = _mesh->get_metric(n0);
+    const double *m0 = _mesh->get_metric(n0);
 
     const real_t *x1 = _mesh->get_coords(n1);
-    const float *m1 = _mesh->get_metric(n1);
+    const double *m1 = _mesh->get_metric(n1);
 
     real_t weight = 1.0/(1.0 + sqrt(property->length(x0, x1, m0)/
                                     property->length(x0, x1, m1)));
